server/requirements.txt
Flask==2.3.3
flask-cors==4.0.0
gunicorn==21.2.0
requests==2.31.0
numpy==1.24.3
python-dotenv==1.0.0
structlog==23.1.0
pytest==7.4.0
pytest-cov==4.1.0
pytest-mock==3.11.1
PyYAML==6.0.1
black==23.7.0
flake8==6.0.0
mypy==1.5.1
sphinx==7.1.2
